fix: resolve cleanup entries 122.C1-C3 and the PR #472 nitpicks
CLEANUP ENTRIES - all three now actually done, not just recorded.

`freminal-orchestrator-protocol` says a cleanup entry "is part of the task --
not a separate task, not a TODO comment in code, not a tracking issue
elsewhere." The activation-time plan wrote "Scheduling: not part of Task 122"
on C1 and C2, silently inverting that default; the close-out inherited it and
filed C3 the same way. The entries had become exactly what the protocol forbids.
C2 in particular was an active agents.md violation left open from activation to
close-out.

- 122.C2: `assert_eq!(name.len(), 1)` in `control_key` -- a panic in production
  input handling that the unwrap/expect lints do not catch -- is replaced by
  `let [name_c] = name.as_bytes() else { return None };`.
- 122.C1: four characterisation tests over the whole `egui::Key` range pinning
  BOTH key-encoding functions, including the `Key::Space` divergence
  (`Ctrl(32)` vs `Ascii(32)`) that makes them non-mergeable, and the coverage
  counts (39 vs 71) so a change to either SET is caught even when no individual
  encoding moved. The functions themselves are unchanged; this entry was always
  about pinning behaviour before anyone merges them.
- 122.C3: the six types and `evaluate_frame_dirty_state` move out of
  `widget.rs` into `frame_dirty.rs`, mirroring `pointer_motion.rs` (122.5a) and
  `frame_drain.rs` (122.11a). Byte-identical apart from visibility, `use` lines
  and the module header; nothing widened past `pub(super)`. `widget.rs` goes
  5,796 -> 5,035, which is 101 lines BELOW its `main` size -- so both god files
  end this task smaller than they began it.

PR #472 NITPICKS - 3 fixed, 1 no-op, 1 declined.

- Misnamed scroll cache: `interface.rs` clamps the request to
  `max_scroll_offset` and forces 0 on the alternate screen BEFORE storing, so
  `previous_requested_scroll_offset` was a lie -- a regression 122.13
  introduced against the original, accurate `previous_scroll_offset`. Renamed
  to `previous_effective_scroll_offset`. CodeRabbit flagged only that field;
  `previous_extra_flatten_rows` has the identical defect and was found by
  inspection -- now `previous_effective_extra_rows`. Both document why they are
  deliberately not `requested`.
- Untested veto terms: added `selection_changed`, `text_blink_changed`,
  `image_pixels_changed` and `image_frame_changed` tests. Mutation-checked --
  neutering `!selection_changed` in the veto fails the new test. `hover_changed`
  is deliberately skipped: the shared `call()` helper hardcodes
  `gutter_inset: 0.0`, which short-circuits `compute_command_block_hover_rows`
  before it reads the mouse position, so driving it needs either a signature
  change touching every existing call site or an elaborate inline fixture.
- Duplicate test pane builders in `frame_drain.rs` merged into one
  `test_pane_with`, with the two originals as thin wrappers. No call site
  changed. -45 lines.
- Drag-release allocation: no change; CodeRabbit states none is required and the
  note records a deliberate cross-file agreement.
- DECLINED, deriving `hover_region_risk` from the already-computed terms: the
  "makes agreement structural" benefit already exists, because
  `pane_hover_region_risk` is DEFINED by calling `pane_hover_region_terms`. What
  remains is three boolean ops evaluated twice per event, which the review
  itself rates low value, and removing it orphans a named concept carrying nine
  tests and four doc references in two other files.

Docs updated to match: the cleanup entries are marked resolved with what each
fix was, the execution state no longer claims they are open, the line-growth
accounting records `widget.rs` at -101 rather than +645, and 122.13's verbatim
name list is corrected for the two `previous_effective_*` renames.

Verified: cargo test --all, cargo test --all --features frame-profiling, both
clippy forms, cargo machete, cargo xtask check-windows.
